add custom rules; update version
